+++ build_resolvers/CHANGELOG.md
@@ -8,6 +8,7 @@
   algorithm, preventing stack overflows.
 - Move `BuildStepImpl` to `build_runner_core`, use `SingleStepReader` directly.
 - Stop building `transitive_digest` files by default.
+- Use `LibraryCycleGraphLoader` to load transitive deps for analysis.
 
 ## 2.4.4