+++ build_runner_core/CHANGELOG.md
@@ -26,6 +26,7 @@
 - New change detection algorithm.
 - Add `reportUnusedAssetsForInput` to `BuildOptions`, to listen for when
   a builder notifies that an asset is unused.
+- Use `LibraryCycleGraphLoader` to load transitive deps for analysis.
 
 ## 8.0.0